CCMSPUI-458 Added test case to EbsApiClientIntegrationTest for search cases
